feature: stripe payment

+++ hiv-and-medical-system/src/main/java/com/swp391_se1866_group2/hiv_and_medical_system/service/dto/request/ServiceCreationRequest.java
@@ -11,6 +11,6 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ServiceCreationRequest {
     String name;
-    double price;
+    Long price;
     ServiceType serviceType;
 }
